- finished minimal version of code to allow for fixing starting points during ffpilot simulation

+++ src/c/lm/tiling/Tiling.cpp
@@ -106,7 +106,7 @@ uint Tiling::getTileIndex(double opVal) const
     return upper - tilingMsg->edges().begin();
 }
 
-void Tiling::setBasin(int basinIndex)
+void Tiling::setBasin(int64_t basinIndex)
 {
     set_current_basin_index(basinIndex);
     if (getTileIndexFromBasin(basinIndex)!=0)
